refactor: changed integer to long

+++ src/main/java/com/example/rebrain/controllers/TestController.java
@@ -40,7 +40,7 @@ public class TestController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<TestDto> getTestById(@PathVariable Integer id) {
+    public ResponseEntity<TestDto> getTestById(@PathVariable Long id) {
         log.debug("Fetching test with ID: {}", id);
         TestDto test = TestMapper.toDto(testService.getOne(id));
         log.debug("Test found: {}", test);
@@ -48,7 +48,7 @@ public class TestController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTest(@PathVariable Integer id) {
+    public ResponseEntity<Void> deleteTest(@PathVariable Long id) {
         log.info("Deleting test with ID: {}", id);
         testService.delete(id);
         log.info("Test with ID: {} deleted", id);
