+++ src/main/java/com/example/rebrain/repositories/CardRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CardRepo extends JpaRepository<CardEntity, Integer> {
+public interface CardRepo extends JpaRepository<CardEntity, Long> {
     List<CardEntity> findByTitle(String title);
 }
